Break particle into its own cpp and header, add simulation file
