+++ CMakeLists.txt
@@ -47,6 +47,8 @@ target_sources(RandomParticleEngine
     src/particle_system.h
     src/controls.h
     src/utils.h
+    src/particle_simulation.h
+    src/particle.h
 
     src/camera.cpp
     src/particle_system.cpp
@@ -54,6 +56,8 @@ target_sources(RandomParticleEngine
     src/main.cpp
     src/shader.cpp
     src/controls.cpp
+    src/particle_simulation.cpp
+    src/particle.cpp
 
 )
 target_include_directories(RandomParticleEngine PRIVATE src)
