+++ src/particle.cpp
@@ -1,0 +1,11 @@
+
+#include <particle.h>
+
+namespace rpg {
+
+Particle::Particle(glm::vec3 pos, glm::vec3 color) {
+  this->pos = pos;
+  this->color = color;
+  this->velocity = glm::vec3{0, 0, 0};
+}
+}  // namespace rpg
