+++ src/particle_simulation.h
@@ -1,0 +1,11 @@
+
+#include <particle.h>
+
+#include <vector>
+namespace rpg::simulation {
+
+const float gravitational_constant = 9.81f;
+
+void simulate_particles(const std::vector<Particle> P);
+
+}  // namespace rpg::simulation
